Add a lame index page for the tutorials. Very lame, needs to look much better, and at least have the contents list generated automatically by the actual documentation constructors. Also note that the package data gets woven into the tutorials, but probably this isn't such a bad thing, since it tells the reader where the examples are.
Added a Felix tangler to interscript, so the tutorial examples
are now have syntax highlighting. The tangler is currently
just a copy of the Python one with the correct keywords for
Felix, so it doesn't get everything right yet.

The tangler is presented as an ordinary pak file and generates
interscript plugins in the interscript directory, therefore
these plugins will not work if the client happens to be
using an installed interscript. In that case they will have
to install the files in lib/python2.4/site-packages/interscript
by hand.

Editors of this pak should note that any error will be
close to catastrophic: the tangler must load correctly even
when not weaving. If this fails, the developer may be left
unable to tangle anything, including a fixed version of the
tangler (this is a typically boostrapping problem).


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@583 3fca3a5c-071c-0410-b500-ba939811c8f2