added poll ts_poll demuxer.
conditionalised demuxer usage, so linux no longer implies epoll,
and solaris no longer implies evtports. this means that older
systems should fall back to select and that we may get the rare
nice surprise of finding a system that supports an unexpected
demuxer (like osx10.4 with poll) or a linux with kqueues. who knows.
ooh, case in point: cygwin supports poll. It spews warnings but
it does work.


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@770 3fca3a5c-071c-0410-b500-ba939811c8f2