Changed serialisation of thread count in worker_fifo to serialise external attempts to change or read the thread count. The prior serialisation wasn't enough to prevent two attempts to change the worker pool size breaking the set_nthreads routine, which assumes the count changes monotonically. This is now enforced by serialising access to that routine entirely. An unfortunate side-effect is that the get_nthreads is also serialised, preventing monitoring of the progress of get_nthreads. This could be solved by using two locks.
The serialisation is not applied to the constructor.
It is not necessary, if the object is constructed on the heap,
and it is no different to accessing an uninitialised object
if it is constructed on the stack.

Serialising the destructor would merely hide a design problem:
the object should not be used once an attempt to destroy it
is made. This can't be enforced without a more advanced
resource manager. A garbage collector would enforce it
(since the object would only be destroyed when it became
unreachable).


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@919 3fca3a5c-071c-0410-b500-ba939811c8f2