Major overhaul of inlining.
* if the argument is an explicit tuple like (1,2) then
each component is passed directly, without going through an
intermediate tuple (smaller faster easier to read code)

* In the case the choice to evaluate eagerly (set a variable),
or lazily (substitute the argument for the parameter), is made
on a parameter by parameter basis, for each val parameter.

* Var parameters are always assigned the argument (eagerly)

* Unchanged: recursive routines always use eager evaluation

* Ref and fun parameters not implemented yet.

* For functions, inline and noinline functions should now have
well defined (and the same!) semantics.

* For procedures, there is still an issue if a val parameter
depends on a variable the procedure changes it should be
inlined eagerly (to simulate call by value). There's a switch
to set the default for vals, but it is always set to Lazy at
the moment.

* Note that even eagerly evaluated vals may be substituted by
the conservative inliner, which uses linear control flow analysis
to detemine if an expressions value can change, however this
routine gives up when it hits a label. A full control flow
analysis is required to make this optimal.


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@922 3fca3a5c-071c-0410-b500-ba939811c8f2