Integrate some C parsing into Felix, initially using
  cparse "kugkjgh";

directive. Probably we need 3 commands here: cface specifies
an interface, but generates no code, cemit emits the code
as written, but also creates Felix bindings to the
interface as in 'cface', and clate translates the C
to Felix. Then we will need versions of the same things
that read whole files.



git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1088 3fca3a5c-071c-0410-b500-ba939811c8f2