Add first part of code designed to convert functions to procedures. Even when this is done, it will only handle the first order case, i.e. direct applications
	f a

where f is a known function. Second order applications,
where f is an arbitrary expression can be done too,
but is much harder. If extended to primitives, this would
allow eliminating ALL functions and function calls from
the program, and thus eliminating the machine stack
(except for holding temporaries.. no return addresses
would be on the stack).

In practice it is evem more complex, because we actually
PREFER the machine stack when it is ok to use it.



git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1700 3fca3a5c-071c-0410-b500-ba939811c8f2