port the build shell scripts to portable python.
The purpose of this patch is to have similar workflows between unix
and windows systems. With this patch, windows now can do out-of-checkout
builds by running "python %FLX_DIR%\autogen.py" in any directory where
FLX_DIR is the source from the repository.

A word of warning, this changes a configuration option called ISCR,
and will cause errors in a prebuilt directory. Instead of having ISCR
be a string to execute iscr, it's just pointing to the executable.
This is needed if python is not specified in the path.

So you don't have to reconfigure, edit config/iscr_config.py and
remove "PYTHONPATH=. python -o " from ISCR line so that it's just
pointing to an executable.

This makes the old build scheme for windows mostly redundant. However,
it's still in place if anyone runs into any bugs.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1717 3fca3a5c-071c-0410-b500-ba939811c8f2