Changed flx optimisation switches: optimisation is on by default with inline=100 now. Use --nooptimise to turn it off.
Removed 'ifnotgoto' statement, added BEXPR_not combinator.
Added branch prediction keywords "likely" and "unlikely".
These express whether a boolean is likely or unlikely
to be true, and thus when used inside a conditional goto,
whether the branch is likely to be taken or not.

This emits a macro which is expanded to a gcc builtin
function if the config detects its availability.
Likely and unlikely have no semantics, they're just
optimisation hints. Tests do not indicate they do anything
on AMD64, but who knows?

These features are used in the parser to ensure loop
exit and loop continuation branchs are marked unlikely
and likely, in the hope this may improve loop performance
(but it doesn't seem to, the only test I did indicates
a slowdown, which makes no sense, because semantically
the gcc builtin has no semantics either: the only
impact shoudl be in reordering instructions to help
processor branch prediction)




git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1776 3fca3a5c-071c-0410-b500-ba939811c8f2