Add new flxg option: --cache_dir=dirname
If specified, Felix will additionally look in directory
dirname for *.par and *.syntax files.

If specified and it exists, and there is a need to write
a cache file, Felix will write in the cache directory.

This allows saving a caches even if the input directory
is read-only, for example, using installed library
with a new version of Felix.

The cache directory doesn't have to exist, but it
will not be created if it doesn't exist. 

Failure to read or write a cache file is never
an error -- it's just an optimisation.
[More precisely, failure during I/O is an error,
but failure to open isn't]

Note that all the cache files are saved in a flat
namespace, which may lead to conflicts (this is a 
bug I'm too lazy to fix at the moment .. there is no
portable way to create directories due to platform
differences in the way permissions are handled).



git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1782 3fca3a5c-071c-0410-b500-ba939811c8f2