added some assoc_list functions, which revealed some new regressions.
First, felix doesn't seem to like:

typedef assoc_list = List::list[(A,B)];

As assoc_list was initially designed. At first it's not a compile
issue, but when actually trying to use it felix complains about a
qualified name lookup error saying list is not a type variable.
Changing it to:

typedef assoc_list = List::list[A*B];

Seems to fix it. Also, there seems to be a code-gen bug that is
revealed by the new regression test I added. Not sure why that's
happening.

Also, I redefined some of the list functions to use other functions
to reduce code duplication. The optimizer should clean it up so
there shouldn't be a performance hit.


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1790 3fca3a5c-071c-0410-b500-ba939811c8f2