Some fiddles to debug the new gc strategy.
The environment variable FLX_ALLOW_COLLECTIONS_ANYWHERE must
be set to enable collection on allocation.

In addition, the compiler passes a context dependent switch
to enable it on each actual collection. These are all currently
set to true.

Collection on allocation is enable if both flags are true.

If enabled, the environment variables FLX_GC_FREQ, FLX_MIN_MEM
determine if a collection might occur. FLX_GC_FREQ is how
often a collection might be attempted, default 1000.

A counter is  incremented on every allocation for which collection 
is enabled (this is probably wrong, it should count all the
allocations, even if collection isn't enabled).

FLX_MIN_MEM is the initial minimum store which must be allocated
before a collection is triggered, default 10Meg. 
After a collection,  this value is reset to the actual 
memory usage multiplied by FLX_FREE_FACTOR (float), 
which defaults to  1.1.

So, to actually STRESS the new system you need to run Felix with
these environment variables:

FLX_ALLOW_COLLECTION_ANYWHERE=1 
FLX_GC_FREQ=1
FLX_MIN_MEM=0

plus any debug switches. This should cause a collection
every allocation (i.e. extremely slow!)

At present this is known to cause problems, probably because
allocated objects are being saved on the stack somewhere,
possibly right at the start of the program. Slightly sloppier
values seem to work in most single threaded cases.

Felix HAS the technology to force everything onto the heap
AND to detect stack code, however there is no way to propagate
the fact that there is stuff on the stack. With gcc and I think
MSVC++ this probably won't be necessary for single threads,
since it is easy enough to find the stack pointer with an
intrinsic, or in more generality a small assembler routine
can find it.

The threading case is much harder. Currently, Felix DOES do
a world stop to collect, even if the collection is triggered
inside an allocation. However this means threads can
be stopped at any allocation, not just when they do service
calls (at which point the stack is empty). It may also be
possible to send threads a signal to force them to stop
(Boehm does that). However none of this stuff is reliable
or portable. Note Felix has an easier job than Boehm,
provided we ban holding Felix pointers outside Felix code,
except for callbacks, where the client data pointer can
be registered as a root.




git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1812 3fca3a5c-071c-0410-b500-ba939811c8f2