Attempt to unravel RTL dependencies. New package flx_rtl_config just contains C++ header files, it is the root of the RTL packages. Exceptions come next, so everyone can use them.
Judy is a leaf, doesn't use any of these.

Then we have threading.

Then we have the garbage collector, which depends on threads
and Judy.

Finally the 'rtl' is all the rest of the architecture.
It should probably be split up more, to make it easier
to understand.

Virgin rebuild is necessary to make all this work.
[One day the system will allow proper cleanup on upgrade .. ]



git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1819 3fca3a5c-071c-0410-b500-ba939811c8f2