Some WIP on gc which currently bugs out FLX_FINALISE, don't use this yet please!
Allow generators to be inlined.

Make read generator inline. This function
MUST be inlined. It cannot be used as a function closure.
[Need to add a 'force_inline' flag, this means that the
generator code is a procedure in disguise]

Fix bug in read function, wrong addressing mode
was used -- passed loc, instead of the address of loc.
This was disguised by the casts involved.



git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1828 3fca3a5c-071c-0410-b500-ba939811c8f2