I think I fixed most of the core dumps, found a couple of stupid bugs. Remaining issue, which probably cannot be solved, is unlinking of dynamically loaded libraries.
This might be effected more safely by a new idea: run
a collection WITHOUT conservative stack scanning
prior to unlinking. If that invalidates some live
pointers the responsibility would be the users.
However it isn't clear how to make that stick without
requiring the user to explicitly unload the library.



git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1829 3fca3a5c-071c-0410-b500-ba939811c8f2