Add xfiles variable to spkgs. If specified it gives a list of files outside interscript control which must be time-stamp checked when considering whether to rebuild or not.
The rtl has an dummy file 'FORCE' listed.
Unless this file is older than the last build time,
the RTL will be rebuilt.

note -- a missing file is assumed to be 'not created
yet' and have a time stamp in the future.

So until you say 'touch FORCE' you will get a rebuild
of rtl every time. Once you touch it you get one more
rebuild for each touch.

This is just a test case for the build system.



git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1840 3fca3a5c-071c-0410-b500-ba939811c8f2