add support for building from checked in code.
This is a substantial change to the build system, and will probably
require a re-configure. The basic idea is that when we build a file,
we first check if that file exists in the checkout directory.

Second, pretty much all the generated files, the object files, lexer
and parser files, and etc, are now stored in a directory called "build",
similar to how ocamlbuild does things. This will ultimately make it
much similar to clean up files by just deleting the build directory.

Third, I moved all the required library and header files into lib/rtl,
so that it would be easier to know what to install. It's still probably
too much.

Finally, to build  tstart_t into the shared pthread header.

Fyi, pthreads on windows don't work, but a good chunk of the tests
succeed. Also, pack files aren't currently supported.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1851 3fca3a5c-071c-0410-b500-ba939811c8f2