Get felix to compile again.
There were a couple bugs in pulling out the rtl code. First, since the
rtl files are in the build directory, we need to make sure they're
referenced through the FLX_DIR variable. Second, the static driver plugin
was putting the output file in the wrong location. Third, there was a
missing comma in an xfiles line.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1871 3fca3a5c-071c-0410-b500-ba939811c8f2