Remove #line numbers and add a unique config.hpp for sockety.
I had also done this in my local copy, but I had a few additional changes.
Skaller mentioned previously that it was a shame to lose the locality that
you get from .pak files. So to recover this, I moved each os-specific
demuxer into it's own subdirectory.

Also, because the demux_sockety.cpp file is the only file to use the
FLX_SOCKLEN_T typedef, I've created a dedicated demux_sockety_config.hpp
that has the generated info.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1893 3fca3a5c-071c-0410-b500-ba939811c8f2