simplify down mkplugins/make.py
There is a lot of commonality between the subphase plugins. This patch
exploits that by just iterating over the function call instead of manually
calling each one. To assist the optional subphases, I lifted things like
HAVE_DYNAMIC_LINKING into the actual subphase. Ultimately, the plan is
to just do away with the "make.py" plugin and have the build system user
provide a list of subphases to build for each phase.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1940 3fca3a5c-071c-0410-b500-ba939811c8f2