Explicitly import the config package everywhere.
This will probably require a reconfiguration.

The old technique of implicitly importing the config files into the global
namespace had some issues with it. The main difficulty is that it forced
us to execfile a good portion of our build system because we needed to
share the same global data. By changing over to an explicitly passed around
config object, we can proceed to greater modularization of the system.
We also will avoid the general issues associated with global variables,
so we'll avoid strange name collision problems. It'll also be clear where
a variable is coming from. This patch should help clean up the code for
eventual separation from the rest of felix.

Also, I renamed config/config.py to config/__init__.py so it's easier to
import. Finally, I did some general import cleanup and removed
mkplugins/ocaml_env.py and mkplugins/link_model_env, which were not being
used.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1942 3fca3a5c-071c-0410-b500-ba939811c8f2