Rework build scripts to be more generic.
This patch is another step on the grand redesign of the build system to
be usable by other projects. The idea is to make a toplevel file the
enduser provides, currently called fbuildroot.py, that right now describes
all the phases the build system can support. This is as opposed to hard
coding the information into script/maker and mkplugins/make.py.

Since the idea of "$PHASE_selected_packages" is now more generic, I had
to modify cal_pkgs.py. For simplicity, I just pulled the functionality
into script/maker since I don't think maker should be explicitly calling
a plugin.

One new feature was added, which was explicitly specifying if one phase
depends on another. This is done by simply adding the phase name to the
fbuild_processes, and that phase will be run before the current phase. This
means that we won't run into cases of stale code if we want to run a test.

One last thing to note is that there's a minor hack I added for the
documentation plugins. Previously, a couple, like rtldoc, were run without
referencing a particular package. This is changed in order to work with
the generic process/phase code. So, instead of just running over and over
for each package, there's a state flag that gets set after the first run.
The code should probably be changed to use the passed in package info to
build docs for just that package.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@1948 3fca3a5c-071c-0410-b500-ba939811c8f2