Make script/maker.py a series of functions instead of all global code.
This is another change that helps us work towards cleaner code. Because
we aren't using execfiles anymore, we don't need all this global data,
even if it's specific to maker.py. So, we'll hide much of the code behind
functions with the plan that later on we'll migrate more of the code out
of the main script and into various libraries.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2002 3fca3a5c-071c-0410-b500-ba939811c8f2