fix build system to work with visual studio 2008 express.
There were a couple problems with cl.exe from 2008. It seems that Microsoft
has made command line parsing a lot more strict. So, while 2005 allowed for
the typo "/GR/MD", 2008 requires there to be actual separation between
the arguments. To fix this, I just unified a lot of the spacing issues so
that the only arguments now that need a space are for gcc, like "-o", but
don't for 2008, with "/OUT:".

There was some other minor cleanups to get 2008 to work. Since 2008 express
does not have the openmp libraries, but still supports "/openmp", I had to
modify the openmp detection. Also, since my checkout directory is on a
read only partition, I changed some of the shutil.copy to shutil.copyfile
so that we don't preserve read only permissions.

These tests still hang, so I manually had to create the stamp to skip them:

pthread-1.01-0.flx
pthread-1.02-0.flx
tut-03.09-0.flx


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2030 3fca3a5c-071c-0410-b500-ba939811c8f2