Apply skaller's concept of mountpoints to get the regression tests to work.
This patch also helps to clean up the build/test directory. Previously,
we had thrown all the test output files into that directory because we
had no concept the mountpoint/relativepath distinction for tests outside
of iscr. Now that we're starting to use the mountpoint/relativepath, we
can put the tests in the proper subdirectory. So, a test named:

test/pthread/pthread-1.01-0.flx

Will have all of it's output files in:

build/test/pthread

Even if that file is not in iscr.

This patch also fixes building the test files with absolute paths. flxg
cannot handle absolute paths, but it can have an absolute path as an
include file. So, to fix this, we just specify the basename of the .flx
file instead of the full path.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2054 3fca3a5c-071c-0410-b500-ba939811c8f2