Extract the build system.
This patch finally pulls most of the build system out of iscr. There are
a couple cheats to make it work, such as adding a version.py file that
has some generated data for configuration. There's still a lot more work to
do to organize the code in a clean manner, however.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2074 3fca3a5c-071c-0410-b500-ba939811c8f2