C parsing hack. Writing C functions such as
	int f(int x) { return x; }

in Felix causes an ambiguity, which adding a trailing ; may
or may not resolve: the above can be interpreted as a 
procedure call something like:

  call int f (int x) { return x; }

because 'int' is not a keyword. This might be fixed but
is non-trivial I think. Instead I have disabled the above
syntax entirely, and enabled this one instead:

 	extern int f(int x) { return x; }

You can also use 'static'. Both extern and static
are now keywords. They have NO significance, except
as punctuation to disambiguate the grammar.
Currently this is ONLY allowed for functions, for
procedures you have to still use

	void f(int x) {}

[this may have to be fixed up the same way :]




git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2099 3fca3a5c-071c-0410-b500-ba939811c8f2