Explicitly specify the process in fbuildroot.
This patch lets us clean up another hack, namely that we were forced
to put all of our processes in fbuild.mkplugins and that each process
had to have the same name as the containing module. By explicitly
specifying the process path, we can finally have processes in different
locations. We can even have them in the same file.


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2101 3fca3a5c-071c-0410-b500-ba939811c8f2