Add a hack to get clean and clean_run to work.
It turns out that using explicit paths breaks the ability to do "mk clean"
because we don't automatically load all the processes into the database.
So, to work around this, I added them to the fbuildroot. Unfortunately,
to actually run these commands now, you have to do "mk --phase=clean_run"
or by doing "mk fbuild.mkplugins.clean_run.clean_run".

We'll have to come up with a better solution to this.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2107 3fca3a5c-071c-0410-b500-ba939811c8f2