Add support for processes to be a function.
This change is backwards compatible with Process based processes by
checking if a process listed in fbuildroot is a class, and if not, treat
it like a function. The only complication is that functions cannot have
the additional metadata like the number of failures. However, we might
want to rethink how we report errors anyway, so this might just be a
short term hack.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2108 3fca3a5c-071c-0410-b500-ba939811c8f2