Restructure make_config.py so everything is in a function.
Previously, since make_config put everything into the global namespace
it was difficult finding what could be factored out. This patch helps
things up so that future merges with fbuild will be simpler.

Note that this patch removes the ability to configure different phases
since generation of the config/__init__.py file depends on HOST_CC and
TARGET_CC existing. We might need to fix this for cross compiling.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2123 3fca3a5c-071c-0410-b500-ba939811c8f2