Add a hack to ocaml packages to just specify all the files.
The pre-this-patch version of the build system didn't handle
.mli without a corresponding .ml file very well, and vice
versa. This is a bit of a hack to get this to work. It does
it by adding "caml_modules" which can reference both .ml and
.mli files and building them in the proper order. I'm using
this so that we can minimize extra stuff in dealing with
third party modules like ocs.


git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2135 3fca3a5c-071c-0410-b500-ba939811c8f2