Put the lexer into it's own dedicated module.
The idea with this is to eventually have the frontend not depend on the
lexer. I'm not sure if we can pull this off because flx_macro depends on
the lexer's token stream. Perhaps we can pull macros and desugaring into
it's own module.

I'm also not sure where to put flx_pdoc. It should probably be put in
a dedicated documentation module.

git-svn-id: https://felix.svn.sourceforge.net/svnroot/felix/felix/trunk@2198 3fca3a5c-071c-0410-b500-ba939811c8f2