Allow static felix exes to use dynlink
Previously, felix would fail to properly
load libraries with dynlink when the file
was statically linked. This was because
statically linked files used the same
mechanism as dynlink to load up the start
functions, and that code was hardcoded to
prevent dynamically loaded code. This patch
fixes this behavior by passing in the exe's
functions to dynlink through a different
mechanism, thus allowing the actual dynamic
code to work.