Modify the build system to include a post configuration step which is used to copy *.fpc files from $HOME/.felix/config into the build config directory, either adding or clobbering the default generated ones. This allows the user to fix up any of the wrong guesses of the configuration. This code is executed every time, and always after interscript package extraction.
I need this code on my Mac to make gmp work since the
configuration assumes gmp{xx} is available on the standard
include and library paths, but it is actually in
directory /opt/local instead.

If anything fails (such as extracting the value of HOME from
the environment) then the process returns without error.