Rewrite flx_filesys to clean up the interface.
This patch removes Flx_filesys.find_file with a series of
functions that search for paths, files, and directories. They
also raise a separate exception Missing_path that allows the
frontend to selectively handle this exception in case it's
recoverable.

It also removes the redundant Flx_parse.lookup_file function.