Initial work on a new felix driver.
This patch starts an experimental new driver for felix.
The idea is to collapse all the drivers into one
executable as well as supporting a REPL. It currently
supports interactively reading felix statements, then
printing out the parsed statements, then the macro
expansion of the statements, and finally the desugared
form.