Add a pass that checks for type variables always used under a pointer. This is called in flxg before the first instantiation pass (the one used for monomorphisation). At the moment this code does nothing other than report candidates for instance merging which would be done by setting that type variable to "void". This should eventually cause a single routine to handle pointers to various types as void*.
The code is not complete because it allows calls to C bindings.
A C function could still dereference a pointer. In that case
the expanded code would fail with a type error. We will have to
exclude this until an attribute or something is added to C
function to tell if a type variable is intensionally polymorphic.

Note that some C functions can be polymorphic, if they have
void* arguments, this promises the function can't dereference
the pointer (at least without a cast).

Finally not that C function arguments can be polymorphic even
if the function being passed is not.