Modify flx_frontend to work on lists of bexes and bsyms instead of one-at-a-time
It turns out a lot of flx_frontend really depends on processing a symbol
at a time, and therefore the original design of dealing with the symbols in
order was inappropriate. This patch changes this by forcing flxc to pass
in the list of desugared bound executables and symbols. This should make
optimization easier since it'll essentially occur a function at a time.