The flx.flx command now appears to work. What's more, some apparent bugs in the Python and bash drivers have been fixed.
However the program doesn't use popen to execute things yet,
and this is essential if we're going to scroll output from
these programs to the console, allowing the user to terminate
the program if it is too slow (or goes into an infinite loop).

This program is temporarily put into the tools directory,
however eventually the binary should be put in the bin
directory and it should become the standard command line
driver for all use. However it might be renamed to say
"flx_cmd" so that it can be invoked with a script which
sets the default install directory, so that the user can
edit that.

Of course, the script itself can be edited, but something
has to actually run it.

With the new program the operation can be fixed up in a way
that is too hard with a bash script. Also new features can
be added including the ability to time-limit a program,
the ability to communicate with the running program
using sockets to monitor it, etc.

We also might get better control of the ability to compile and
link multiple files, including C/C++ files.

Also, since Felix can now generate CPython modules as shared libraries
we might want to run a Python script using some generated modules.

There's also some possibility of launching a GUI and maybe
managing some kind of interpretive looping thing (but don't
hold your breath Erick .. :)

It remains to do some more testing, and then put this thing where
it belongs. Also flx_maker.pak has too much crud in it,
this program does require configuration information however.