Implement composition. A bit of a hack. Given the expression   f . g
Felix first tries reverse appliction g (f). If that fails and
f and g are functions, it generates a reverse composition
such that (f. g) x = g ( f(x)). This means that operator . appears
to be associative:

  x . f . g = (x . f) . g = x . (f . g)

It is really a hack. It is unlikely the above would be ambigious
due to overloading, but it is possible. In this case note again
application is tried first. For example given

fold f i ls

you can write

f . fold i ls

instead. This isn't a composition, it's an application of fold to f.

You should note that

  x . f = {x} . f ()

In other words, if we lift a value x to a function k()=> x, then we
can actually compose it with f:

  {x}.f

is a the composition which when applied to () will apply f to x.

Felix lets you write x.f instead, and calls it reverse application
but actually we should really consider . as composition,
and, if we get a function accepting unit and returning a T,
when we wanted a T, we just apply it to () to get the T.
Unfortunately, "expecting a T" is an inherited attribute which
conflicts with overloading (the overload would have to be resolved
first to know what is expected but will fail because there's no match).

The implemented hackery reverses the choice order and this
only requires synthetic attributes, so it is easy to implement.