Get the composite stuff working for closures properly. Did this with a hack, fiddled with flx_mkcls to generate the closures (without adding an Heap property).
Actually, this lambda lifting stuff could ALL be done
straight after binding before optimisations. It would
be a bit inefficient, since it might process functions
that never get called.

However it would ensure that every BEXPR_closure was actually
a closure of a Felix function, not a composition, constructor,
C primitive, or whatever: those things would always occur
in applications.

In turn, this would make it easier to add a lot of
new terms and easily relegate closures of them to wrappers.

This is a general problem: many terms take expression arguments,
but can't be used in isolation, lambda lifting fixes this
for any term you care to name.

Doing this lambda lifting then *relies* on the optimiser
to eliminate the closures when it can.