Changed textio to use binary mode and fread instead of text mode and fgets when loading a file, and made sure the string grabbed from the buffer didn't stop at any embedded NUL char. The old code didn't load *.png files properly for the webserver.
Changed the webserver to use the load function instead of
using a stream copy, so that we can get the length of the
file and put that in the HTML header. With this mod,
firefox and curl terminate when the expected amount of
text is sent, instead of waiting for the socket to close.
At least on localhost .. it may not help on the remote server.

The syntax colouring stuff doesn't do this yet, it's a major
slowdown to have to buffer the contents of a dynamically generated
page in memory, just to get the length.. but we may HAVE to do it.