Add the pythonic operator "in", I didn't think this would work because of "let" .. "in" .. but it seems to parse OK in simple cases. We'll see. It reads better than "isin". The parser maps this operator so that x in y becomes mem(x,y). Modified array and list, so you can now say, for example x in (1,2,3,4). Operators "in" currently has the same precedence as a comparison.
It is tempting to define, for example (2,3) < (1,2,3,4) meaning
that for all x in LHA, x is in RHA, i.e. a subset. Unfortunately
there are two other interpretations: it is a subslice, and,
it is a subsequence (a subsequence is any selection of values
in the same order, a subslice is a contiguous subsequence).
Of course none of that matters for sets.

I should add a "mem" function to varray and avl tree, as well
as the STL containers (and Judy arrays). Actually, I should add some more
data structures.. :)