Fix a bug in the compilers handling of calls to C functions. Fiddle unsuccessfully with the failure of the interscript felix.py tangler to load.
Add a new data structure to the library, darray, which is
an unbounded variable length array.

Hack the implementation to work around the restriction that
some operators such as += cause the parse to silently take the
address of the first argument, which, combined with the restriction
that you can only address simple variables, makes it impossible
to say a.b += 1 where a is a struct and b a field of it.
Assignments work, probably because I gave up trying to enforce
the restriction and let the C++ compiler handle it.

In this case, incrementing an "rvalue" is just fine because a varray is
just a pointer, and the increment applies to what it points to.

I need to document the "lvalue" attribute of primitive functions too,
this specifies the function returns an lvalue so that things like

f ( a) = 1;

can work. [This is not the same "lvalue" as the defunct type
notation lvalue[t] which is current typedefed to just t]