Restructure symbol tables a bit. Add debugging functions to display tables. Temporarily the per-file cached symbol tables are being displayed so I can track the next step: merging and binding the standard library (and caching that).
Also changes the print routine for bound types to optionally
use the bound symbol table: only one constructor (BTYP_inst)
actually needs this table. If the table isn't supplied,
just print the primitive index. This makes it possible to
display the bound types in entry_set_t's which are used in
the unbound symbol table to create views.