fix flx.flx, but might cause an fd leak
A process opened by popen has to be closed by pclose, but load_file
automatically called fclose on the file when it was finished. This
caused a double free. This patch removes the fclose, but this may result
in some file descriptor leakage if anything was relying on load_file
automatically closing the file.