Fixup the build for mk_daemon, and add build for timeout. Also copy the test suite to the install, again, so we can browse it. Change popen call in webserver (fdoc2html module) so the @sh command uses timeout. Unfortunately, timeout has to be installed in PATH for this to work as written: probably should change it to use the install directory to find the executable.
The sh command still needs to "htmlize" the output to fixup <>& etc.
This is really just a test case, now I have to design how to check
@felix code in fdoc's for correctness. Eventually this will replace
the old tutorial. At the moment I could do

@felix filename
generates
<pre>
@sh flx filename
</pre>

to present some code and its output: that would include any errors
generated by the flx command, provided the return code is ignored
(which it isn't at the moment, it's used to emit a "bug running
the command" text, while here we really want to see the compiler
diagnostics).