Move interscript from a standalone directory into buildsystem.
The idea is to first do this, then remove parts of it such as all the
weaving code which is no longer needed. Eventually, it can be simplified
enough to proces files such as lpsrc/flx_rtl_config.pak, which would
otherwise be a nightmare to integrate directly into the build system
using fbuild's native attributes.

Next, we try to generalise the support as follows: the idea is that
a client of fbuild can use the mini-interscript as their standard
code generator: this is much more powerful than using *.*.in files
and just replacing scraps between @..@ marks. The key concept here
is to *retain* the idea that these files can use "arbitrary"
configuration data names, and the build system do what is done now:
convert fbuild's attributes into ones that the end user's files
need.

If this pans out, then the mini-iscr can be made a builtin part
of fbuild instead of part of the Felix buildsystem. And in particular
it can be imported to provide support, instead of executed with
the system() style command.