Fix location of html files so they're relative to the --root.
The web is a mess. Firefox misparses requests so it is impossible to
specify an absolute filename. If you put http://.../name it's relative
to the server root (inline links, just say name). If you put instead
http://..//name its the same, and /// gives you // so there's no way
to get a single /. That's why I added the $, to mean "don't search
for this file, use the given name after the $", because once you're
inside, say, and flx file or cpp file some filenames have to be
searched for on a path (eg include files).

Unfortunately this leads to a real mess. The webserver inserts $ into
links it generates but can't see <a href.. > type of links. Grrr.
Anyhow, htmls files were just searched for as they came, and in
particular is you just put the http://website/ URL that gives a
blank filename which is replaced by index.html which causes the
webserver to look in the current directory. Fixed, it now looks
in the specified root.